TST: move custom business month tests to own file (#27085) (#42284)
